Move spec files to their respective folders
