+++ playwright.config.ts
@@ -10,7 +10,7 @@ import { defineConfig, devices } from '@playwright/test';
  * See https://playwright.dev/docs/test-configuration.
  */
 export default defineConfig({
-  testDir: './e2e',
+  testDir: './e2e/spec/',
   /* Maximum time one test can run for. */
   timeout: 30 * 1000,
   expect: {
